Tambah Expression

+++ Expression/AddExpression.cpp
@@ -3,6 +3,6 @@
 AddExpression::AddExpression(Expression *x, Expression *y): BinaryExpression(x, y) {
 }
 
-int AddExpression::solve() {
+float AddExpression::solve() {
 	return (x->solve() + y->solve());
 }
