+++ Expression/AddExpression.hpp
@@ -5,7 +5,7 @@
 class AddExpression: public BinaryExpression {
 public:
     AddExpression(Expression *x, Expression *y);
-    int solve();
+    float solve();
 };
 
 #endif
