+++ Expression/SubstractExpression.cpp
@@ -3,6 +3,6 @@
 SubstractExpression::SubstractExpression(Expression *x, Expression *y): BinaryExpression(x, y) {
 }
 
-int SubstractExpression::solve() {
+float SubstractExpression::solve() {
 	return (x->solve() - y->solve());
 }
